ssh-add side of destination constraints
Have ssh-add accept a list of "destination constraints" that allow
restricting where keys may be used in conjunction with a ssh-agent/ssh
that supports session ID/hostkey binding.

Constraints are specified as "[user]@[host-pattern]". The host pattern
is used to obtain a list of host keys for that destination. The user/
hostkeys are encoded using a new restrict-destination@openssh.com
key constraint for passing to the agent.

The username or the host-pattern is optional. Not specifying a host
will restrict the key to permit signing userauth requests for the
specified username across all hosts. It is also possible to specify
a wildcard host-pattern that will expand to all hosts listed in
known_hosts.

host keys are looked up in the default client user/system known_hosts
files. It is possible to override this set on the command-line.