Implement parsing and use of Bloom filters in KRLs
Bloom filter extensions are parsed and consulted prior to searching
the RB trees, which are only searched if no corresponding Bloom filter
was in the KRL or the Bloom filter returned a match.

TODO: rewrite is_key_revoked() to avoid needing to parse non-Bloom filter
sections at all unless a Bloom filter matched (or was not present).
Currently there is little benefit to the Bloom filter extensions as the
whole KRL needs to be loaded and parsed, which is quite expensive for
large KRLs.