Trim unneccessary code from sshd binary
The sshd binary now only needs to listen, accept connections and reexec
the sshd-monitor binary. As such, most of the code in it can go away.

Some subtleties are presented by the auth2_methods_valid() call - we want
to call this in the sshd binary to validate the configuration, but we
don't want to pull in all of the auth-*.c code. This refactors Authmethod
to use a separate configuration structure that only contains authentication
method names and enabled flags to avoid this.