Improve OpenSSH performance on high BDP links
The SSH protocol performs flow control using a channel receive
window that may only be set at the time a channel is created and
cannot be subsequently modified.

The current default window size is sufficient for most uses, but
on high bandwidth x delay links it is not enough. Increasing the
default window size is possible, but carries the cost of additional
worst-case memory usage if the output side is slow.

This implements a protocol extension to vary the maximum channel
window after the channel has been created, and uses this extension
to grow the window if the TCP receive window increases in size.
TCP has had a half-century of optimisation, so we should copy
its homework wherever possible.

This is based on work by Denzel Strauss and Sally Sun during their
STEP iternship at Google.