Remove use of libcrypto *_set_meth() for PKCS#11
This removes the internal indirection of libcrypto operations for PKCS#11
keys, e.g. RSA_set_meth().

Instead of making EVP_PKEY objects that magically redirect via PKCS#11,
this explicits passes signing operations down to ssh-pkcs11.c in a manner
similar to what we do for SK keys.

This removes a lot of scary pointer chasing and reference counting in
ssh-pkcs11.c and ssh-pkcs11-client.c at the cost of a bit more
complexity in the actual PKCS#11 signing code. This could be simplified
further.